refactor(compiler): move `invalidateCache` to cache module
